[Charles/Emily] fixes theme edit bug.
when the user clicked edit to edit a theme, then made the theme name blank
and tried to save, they would be redirected to the edit page showing theme
field including the region selection dropdown. To fix this issue, we added
javascript validation of the form so that the user is not redirected as
long as the form has invalid input.